split the error classification
